requirements.txt
# requirements.txt
streamlit
python-dotenv
plotly